update lastMessageReadAt and move delayUntilPingInterval into a higher level class
